Rename client folder names
