+++ src/pages/SelectClient/index.tsx
@@ -7,14 +7,14 @@ import { WorkflowPageTemplate } from '../../components/WorkflowPage/WorkflowPage
 import { routeToCorrectWorkflowStep } from '../../utils/RouteToCorrectWorkflowStep';
 import SelectClientSection from './SelectClientSection';
 import SelectClientButtons from './SelectClientButtons';
-import { PrysmDetails } from '../Clients/Eth2/Prysm';
-import { LighthouseDetails } from '../Clients/Eth2/Lighthouse';
-import { TekuDetails } from '../Clients/Eth2/Teku';
-import { NimbusDetails } from '../Clients/Eth2/Nimbus';
-import { GethDetails } from '../Clients/Eth1/Geth';
-import { OpenEthereumDetails } from '../Clients/Eth1/OpenEthereum';
-import { BesuDetails } from '../Clients/Eth1/Besu';
-import { NethermindDetails } from '../Clients/Eth1/Nethermind';
+import { PrysmDetails } from '../Clients/Consensus/Prysm';
+import { LighthouseDetails } from '../Clients/Consensus/Lighthouse';
+import { TekuDetails } from '../Clients/Consensus/Teku';
+import { NimbusDetails } from '../Clients/Consensus/Nimbus';
+import { GethDetails } from '../Clients/Execution/Geth';
+import { OpenEthereumDetails } from '../Clients/Execution/OpenEthereum';
+import { BesuDetails } from '../Clients/Execution/Besu';
+import { NethermindDetails } from '../Clients/Execution/Nethermind';
 import PrysmaticCircle from '../../static/prysmatic-labs-circle.png';
 import LighthouseCircle from '../../static/lighthouse-circle.png';
 import NimbusCircle from '../../static/nimbus-circle.png';
